feat: lock versions with npm overrides
