+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/TaskUpdatePackages.java
@@ -101,7 +101,8 @@ public class TaskUpdatePackages extends NodeUpdater {
                     .getPackages();
             JsonObject packageJson = getPackageJson();
             modified = updatePackageJsonDependencies(packageJson,
-                    scannedApplicationDependencies);
+                    scannedApplicationDependencies)
+                        | lockVersionForNpm(packageJson);
 
             if (modified) {
                 writePackageFile(packageJson);
@@ -122,6 +123,39 @@ public class TaskUpdatePackages extends NodeUpdater {
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
+    }
+
+    private boolean lockVersionForNpm(JsonObject packageJson) throws IOException {
+        if(enablePnpm) {
+            return false;
+        }
+        boolean modified = false;
+        
+        final String versionsPath = generateVersionsJson();
+        
+        File generatedVersionsFile = new File(npmFolder, versionsPath);
+        final JsonObject versionsJson = Json.parse(FileUtils.readFileToString(
+                generatedVersionsFile, StandardCharsets.UTF_8));
+        
+        if (versionsJson != null) {
+            JsonObject overridesSection = getOverridesSection(packageJson);
+            for(String dependency: versionsJson.keys()) {
+                if(!overridesSection.hasKey(dependency)) {
+                    overridesSection.put(dependency, "$"+dependency);
+                    modified = true;
+                }
+            }
+        }
+        return modified;
+    }
+
+    private JsonObject getOverridesSection(JsonObject packageJson) {
+        JsonObject overridesSection = packageJson.getObject(OVERRIEDS);
+        if(overridesSection == null) {
+            overridesSection = Json.createObject();
+            packageJson.put(OVERRIEDS, overridesSection);
+        }
+        return overridesSection;
     }
 
     @Override
